+++ planner/cardinality/BUILD.bazel
@@ -5,6 +5,8 @@ go_library(
     srcs = [
         "histogram.go",
         "pseudo.go",
+        "row_count_column.go",
+        "row_count_index.go",
         "selectivity.go",
         "trace.go",
     ],
@@ -15,6 +17,7 @@ go_library(
         "//kv",
         "//parser/ast",
         "//parser/format",
+        "//parser/model",
         "//parser/mysql",
         "//planner/util",
         "//planner/util/debugtrace",
@@ -24,10 +27,14 @@ go_library(
         "//types",
         "//types/parser_driver",
         "//util/chunk",
+        "//util/codec",
+        "//util/collate",
         "//util/logutil",
+        "//util/mathutil",
         "//util/ranger",
         "//util/tracing",
         "@com_github_pingcap_errors//:errors",
+        "@com_github_pingcap_failpoint//:failpoint",
         "@org_golang_x_exp//maps",
         "@org_uber_go_zap//:zap",
     ],
@@ -38,13 +45,14 @@ go_test(
     timeout = "short",
     srcs = [
         "main_test.go",
+        "row_count_test.go",
         "selectivity_test.go",
         "trace_test.go",
     ],
     data = glob(["testdata/**"]),
     embed = [":cardinality"],
     flaky = True,
-    shard_count = 29,
+    shard_count = 30,
     deps = [
         "//config",
         "//domain",
