+++ planner/cardinality/trace.go
@@ -16,15 +16,21 @@ package cardinality
 
 import (
 	"bytes"
+	"encoding/json"
 	"errors"
 
 	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/parser/ast"
 	"github.com/pingcap/tidb/parser/format"
+	"github.com/pingcap/tidb/parser/model"
+	"github.com/pingcap/tidb/planner/util/debugtrace"
 	"github.com/pingcap/tidb/sessionctx"
+	"github.com/pingcap/tidb/sessionctx/stmtctx"
+	"github.com/pingcap/tidb/statistics"
 	driver "github.com/pingcap/tidb/types/parser_driver"
 	"github.com/pingcap/tidb/util/chunk"
 	"github.com/pingcap/tidb/util/logutil"
+	"github.com/pingcap/tidb/util/ranger"
 	"github.com/pingcap/tidb/util/tracing"
 	"go.uber.org/zap"
 )
@@ -107,4 +113,194 @@ func exprToString(e expression.Expression) (string, error) {
 		return buffer.String(), nil
 	}
 	return "", errors.New("unexpected type of Expression")
+}
+
+/*
+ Below is debug trace for GetRowCountByXXX().
+*/
+
+type getRowCountInput struct {
+	Ranges []string
+	ID     int64
+}
+
+func debugTraceGetRowCountInput(
+	s sessionctx.Context,
+	id int64,
+	ranges ranger.Ranges,
+) {
+	root := debugtrace.GetOrInitDebugTraceRoot(s)
+	newCtx := &getRowCountInput{
+		ID:     id,
+		Ranges: make([]string, len(ranges)),
+	}
+	for i, r := range ranges {
+		newCtx.Ranges[i] = r.String()
+	}
+	root.AppendStepToCurrentContext(newCtx)
+}
+
+// GetTblInfoForUsedStatsByPhysicalID get table name, partition name and TableInfo that will be used to record used stats.
+var GetTblInfoForUsedStatsByPhysicalID func(sctx sessionctx.Context, id int64) (fullName string, tblInfo *model.TableInfo)
+
+// recordUsedItemStatsStatus only records un-FullLoad item load status during user query
+func recordUsedItemStatsStatus(sctx sessionctx.Context, stats interface{}, tableID, id int64) {
+	// Sometimes we try to use stats on _tidb_rowid (id == -1), which must be empty, we ignore this case here.
+	if id <= 0 {
+		return
+	}
+	var isIndex, missing bool
+	var loadStatus *statistics.StatsLoadedStatus
+	switch x := stats.(type) {
+	case *statistics.Column:
+		isIndex = false
+		if x == nil {
+			missing = true
+		} else {
+			loadStatus = &x.StatsLoadedStatus
+		}
+	case *statistics.Index:
+		isIndex = true
+		if x == nil {
+			missing = true
+		} else {
+			loadStatus = &x.StatsLoadedStatus
+		}
+	}
+
+	// no need to record
+	if !missing && loadStatus.IsFullLoad() {
+		return
+	}
+
+	// need to record
+	statsRecord := sctx.GetSessionVars().StmtCtx.GetUsedStatsInfo(true)
+	if statsRecord[tableID] == nil {
+		name, tblInfo := GetTblInfoForUsedStatsByPhysicalID(sctx, tableID)
+		statsRecord[tableID] = &stmtctx.UsedStatsInfoForTable{
+			Name:    name,
+			TblInfo: tblInfo,
+		}
+	}
+	recordForTbl := statsRecord[tableID]
+
+	var recordForColOrIdx map[int64]string
+	if isIndex {
+		if recordForTbl.IndexStatsLoadStatus == nil {
+			recordForTbl.IndexStatsLoadStatus = make(map[int64]string, 1)
+		}
+		recordForColOrIdx = recordForTbl.IndexStatsLoadStatus
+	} else {
+		if recordForTbl.ColumnStatsLoadStatus == nil {
+			recordForTbl.ColumnStatsLoadStatus = make(map[int64]string, 1)
+		}
+		recordForColOrIdx = recordForTbl.ColumnStatsLoadStatus
+	}
+
+	if missing {
+		recordForColOrIdx[id] = "missing"
+		return
+	}
+	recordForColOrIdx[id] = loadStatus.StatusToString()
+}
+
+// ceTraceRange appends a list of ranges and related information into CE trace
+func ceTraceRange(sctx sessionctx.Context, tableID int64, colNames []string, ranges []*ranger.Range, tp string, rowCount uint64) {
+	sc := sctx.GetSessionVars().StmtCtx
+	allPoint := true
+	for _, ran := range ranges {
+		if !ran.IsPointNullable(sctx) {
+			allPoint = false
+			break
+		}
+	}
+	if allPoint {
+		tp = tp + "-Point"
+	} else {
+		tp = tp + "-Range"
+	}
+	expr, err := ranger.RangesToString(sc, ranges, colNames)
+	if err != nil {
+		logutil.BgLogger().Debug("Failed to trace CE of ranges", zap.String("category", "OptimizerTrace"), zap.Error(err))
+	}
+	// We don't need to record meaningless expressions.
+	if expr == "" || expr == "true" || expr == "false" {
+		return
+	}
+	ceRecord := tracing.CETraceRecord{
+		TableID:  tableID,
+		Type:     tp,
+		Expr:     expr,
+		RowCount: rowCount,
+	}
+	sc.OptimizerCETrace = append(sc.OptimizerCETrace, &ceRecord)
+}
+
+/*
+ Below is debug trace for the estimation for each single range inside GetRowCountByXXX().
+*/
+
+type startEstimateRangeInfo struct {
+	Range            string
+	LowValueEncoded  []byte
+	HighValueEncoded []byte
+	CurrentRowCount  float64
+}
+
+func debugTraceStartEstimateRange(
+	s sessionctx.Context,
+	r *ranger.Range,
+	lowBytes, highBytes []byte,
+	currentCount float64,
+) {
+	root := debugtrace.GetOrInitDebugTraceRoot(s)
+	traceInfo := &startEstimateRangeInfo{
+		CurrentRowCount:  currentCount,
+		Range:            r.String(),
+		LowValueEncoded:  lowBytes,
+		HighValueEncoded: highBytes,
+	}
+	root.AppendStepWithNameToCurrentContext(traceInfo, "Start estimate range")
+}
+
+type debugTraceAddRowCountType int8
+
+const (
+	debugTraceUnknownTypeAddRowCount debugTraceAddRowCountType = iota
+	debugTraceImpossible
+	debugTraceUniquePoint
+	debugTracePoint
+	debugTraceRange
+	debugTraceVer1SmallRange
+)
+
+var addRowCountTypeToString = map[debugTraceAddRowCountType]string{
+	debugTraceUnknownTypeAddRowCount: "Unknown",
+	debugTraceImpossible:             "Impossible",
+	debugTraceUniquePoint:            "Unique point",
+	debugTracePoint:                  "Point",
+	debugTraceRange:                  "Range",
+	debugTraceVer1SmallRange:         "Small range in ver1 stats",
+}
+
+func (d debugTraceAddRowCountType) MarshalJSON() ([]byte, error) {
+	return json.Marshal(addRowCountTypeToString[d])
+}
+
+type endEstimateRangeInfo struct {
+	RowCount float64
+	Type     debugTraceAddRowCountType
+}
+
+func debugTraceEndEstimateRange(
+	s sessionctx.Context,
+	count float64,
+	addType debugTraceAddRowCountType,
+) {
+	root := debugtrace.GetOrInitDebugTraceRoot(s)
+	traceInfo := &endEstimateRangeInfo{
+		RowCount: count,
+		Type:     addType,
+	}
+	root.AppendStepWithNameToCurrentContext(traceInfo, "End estimate range")
 }
