+++ statistics/BUILD.bazel
@@ -48,7 +48,6 @@ go_library(
         "//util/memory",
         "//util/ranger",
         "//util/sqlexec",
-        "//util/tracing",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_pingcap_tipb//go-tipb",
@@ -76,7 +75,7 @@ go_test(
     data = glob(["testdata/**"]),
     embed = [":statistics"],
     flaky = True,
-    shard_count = 42,
+    shard_count = 41,
     deps = [
         "//config",
         "//parser/ast",
