+++ statistics/handle/handletest/BUILD.bazel
@@ -14,6 +14,7 @@ go_test(
         "//config",
         "//domain",
         "//parser/model",
+        "//planner/cardinality",
         "//session",
         "//sessionctx/variable",
         "//statistics",
