+++ statistics/histogram.go
@@ -382,10 +382,10 @@ func (hg *Histogram) ToString(idxCols int) string {
 	return strings.Join(strs, "\n")
 }
 
-// equalRowCount estimates the row count where the column equals to value.
+// EqualRowCount estimates the row count where the column equals to value.
 // matched: return true if this returned row count is from Bucket.Repeat or bucket NDV, which is more accurate than if not.
 // The input sctx is just for debug trace, you can pass nil safely if that's not needed.
-func (hg *Histogram) equalRowCount(sctx sessionctx.Context, value types.Datum, hasBucketNDV bool) (count float64, matched bool) {
+func (hg *Histogram) EqualRowCount(sctx sessionctx.Context, value types.Datum, hasBucketNDV bool) (count float64, matched bool) {
 	if sctx != nil && sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
 		debugtrace.EnterContextCommon(sctx)
 		defer func() {
@@ -393,12 +393,12 @@ func (hg *Histogram) equalRowCount(sctx sessionctx.Context, value types.Datum, h
 			debugtrace.LeaveContextCommon(sctx)
 		}()
 	}
-	_, bucketIdx, inBucket, match := hg.locateBucket(sctx, value)
+	_, bucketIdx, inBucket, match := hg.LocateBucket(sctx, value)
 	if !inBucket {
 		return 0, false
 	}
 	if sctx != nil && sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
-		debugTraceBuckets(sctx, hg, []int{bucketIdx})
+		DebugTraceBuckets(sctx, hg, []int{bucketIdx})
 	}
 	if match {
 		return float64(hg.Buckets[bucketIdx].Repeat), true
@@ -406,18 +406,18 @@ func (hg *Histogram) equalRowCount(sctx sessionctx.Context, value types.Datum, h
 	if hasBucketNDV && hg.Buckets[bucketIdx].NDV > 1 {
 		return float64(hg.BucketCount(bucketIdx)-hg.Buckets[bucketIdx].Repeat) / float64(hg.Buckets[bucketIdx].NDV-1), true
 	}
-	return hg.notNullCount() / float64(hg.NDV), false
+	return hg.NotNullCount() / float64(hg.NDV), false
 }
 
-// greaterRowCount estimates the row count where the column greater than value.
+// GreaterRowCount estimates the row count where the column greater than value.
 // It's deprecated. Only used for test.
-func (hg *Histogram) greaterRowCount(value types.Datum) float64 {
-	histRowCount, _ := hg.equalRowCount(nil, value, false)
-	gtCount := hg.notNullCount() - hg.lessRowCount(nil, value) - histRowCount
+func (hg *Histogram) GreaterRowCount(value types.Datum) float64 {
+	histRowCount, _ := hg.EqualRowCount(nil, value, false)
+	gtCount := hg.NotNullCount() - hg.LessRowCount(nil, value) - histRowCount
 	return math.Max(0, gtCount)
 }
 
-// locateBucket locates where a value falls in the range of the Histogram.
+// LocateBucket locates where a value falls in the range of the Histogram.
 // The input sctx is just for debug trace, you can pass nil safely if that's not needed.
 //
 // Return value:
@@ -430,12 +430,12 @@ func (hg *Histogram) greaterRowCount(value types.Datum) float64 {
 //
 // Examples:
 // val0 |<-[bkt0]->| |<-[bkt1]->val1(last value)| val2 |<--val3--[bkt2]->| |<-[bkt3]->| val4
-// locateBucket(val0): false, 0, false, false
-// locateBucket(val1): false, 1, true, true
-// locateBucket(val2): false, 2, false, false
-// locateBucket(val3): false, 2, true, false
-// locateBucket(val4): true, 3, false, false
-func (hg *Histogram) locateBucket(sctx sessionctx.Context, value types.Datum) (exceed bool, bucketIdx int, inBucket, matchLastValue bool) {
+// LocateBucket(val0): false, 0, false, false
+// LocateBucket(val1): false, 1, true, true
+// LocateBucket(val2): false, 2, false, false
+// LocateBucket(val3): false, 2, true, false
+// LocateBucket(val4): true, 3, false, false
+func (hg *Histogram) LocateBucket(sctx sessionctx.Context, value types.Datum) (exceed bool, bucketIdx int, inBucket, matchLastValue bool) {
 	if sctx != nil && sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
 		defer func() {
 			debugTraceLocateBucket(sctx, &value, exceed, bucketIdx, inBucket, matchLastValue)
@@ -480,12 +480,12 @@ func (hg *Histogram) LessRowCountWithBktIdx(sctx sessionctx.Context, value types
 	if hg.Bounds.NumRows() == 0 {
 		return 0, 0
 	}
-	exceed, bucketIdx, inBucket, match := hg.locateBucket(sctx, value)
+	exceed, bucketIdx, inBucket, match := hg.LocateBucket(sctx, value)
 	if exceed {
-		return hg.notNullCount(), hg.Len() - 1
+		return hg.NotNullCount(), hg.Len() - 1
 	}
 	if sctx != nil && sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
-		debugTraceBuckets(sctx, hg, []int{bucketIdx - 1, bucketIdx})
+		DebugTraceBuckets(sctx, hg, []int{bucketIdx - 1, bucketIdx})
 	}
 	preCount := float64(0)
 	if bucketIdx > 0 {
@@ -501,8 +501,9 @@ func (hg *Histogram) LessRowCountWithBktIdx(sctx sessionctx.Context, value types
 	return preCount + hg.calcFraction(bucketIdx, &value)*(curCount-curRepeat-preCount), bucketIdx
 }
 
+// LessRowCount estimates the row count where the column less than value.
 // The input sctx is just for debug trace, you can pass nil safely if that's not needed.
-func (hg *Histogram) lessRowCount(sctx sessionctx.Context, value types.Datum) float64 {
+func (hg *Histogram) LessRowCount(sctx sessionctx.Context, value types.Datum) float64 {
 	result, _ := hg.LessRowCountWithBktIdx(sctx, value)
 	return result
 }
@@ -510,27 +511,27 @@ func (hg *Histogram) lessRowCount(sctx sessionctx.Context, value types.Datum) fl
 // BetweenRowCount estimates the row count where column greater or equal to a and less than b.
 // The input sctx is just for debug trace, you can pass nil safely if that's not needed.
 func (hg *Histogram) BetweenRowCount(sctx sessionctx.Context, a, b types.Datum) float64 {
-	lessCountA := hg.lessRowCount(sctx, a)
-	lessCountB := hg.lessRowCount(sctx, b)
+	lessCountA := hg.LessRowCount(sctx, a)
+	lessCountB := hg.LessRowCount(sctx, b)
 	// If lessCountA is not less than lessCountB, it may be that they fall to the same bucket and we cannot estimate
 	// the fraction, so we use `totalCount / NDV` to estimate the row count, but the result should not greater than
-	// lessCountB or notNullCount-lessCountA.
+	// lessCountB or NotNullCount-lessCountA.
 	if lessCountA >= lessCountB && hg.NDV > 0 {
-		result := math.Min(lessCountB, hg.notNullCount()-lessCountA)
-		return math.Min(result, hg.notNullCount()/float64(hg.NDV))
+		result := math.Min(lessCountB, hg.NotNullCount()-lessCountA)
+		return math.Min(result, hg.NotNullCount()/float64(hg.NDV))
 	}
 	return lessCountB - lessCountA
 }
 
 // TotalRowCount returns the total count of this histogram.
 func (hg *Histogram) TotalRowCount() float64 {
-	return hg.notNullCount() + float64(hg.NullCount)
+	return hg.NotNullCount() + float64(hg.NullCount)
 }
 
-// notNullCount indicates the count of non-null values in column histogram and single-column index histogram,
+// NotNullCount indicates the count of non-null values in column histogram and single-column index histogram,
 // for multi-column index histogram, since we cannot define null for the row, we treat all rows as non-null, that means,
-// notNullCount would return same value as TotalRowCount for multi-column index histograms.
-func (hg *Histogram) notNullCount() float64 {
+// NotNullCount would return same value as TotalRowCount for multi-column index histograms.
+func (hg *Histogram) NotNullCount() float64 {
 	if hg.Len() == 0 {
 		return 0
 	}
@@ -814,13 +815,14 @@ func MergeHistograms(sc *stmtctx.StatementContext, lh *Histogram, rh *Histogram,
 // AvgCountPerNotNullValue gets the average row count per value by the data of histogram.
 func (hg *Histogram) AvgCountPerNotNullValue(totalCount int64) float64 {
 	factor := hg.GetIncreaseFactor(totalCount)
-	totalNotNull := hg.notNullCount() * factor
+	totalNotNull := hg.NotNullCount() * factor
 	curNDV := float64(hg.NDV) * factor
 	curNDV = math.Max(curNDV, 1)
 	return totalNotNull / curNDV
 }
 
-func (hg *Histogram) outOfRange(val types.Datum) bool {
+// OutOfRange checks if the datum is out of range for this Histogram.
+func (hg *Histogram) OutOfRange(val types.Datum) bool {
 	if hg.Len() == 0 {
 		return false
 	}
@@ -828,7 +830,7 @@ func (hg *Histogram) outOfRange(val types.Datum) bool {
 		chunk.Compare(hg.Bounds.GetRow(hg.Bounds.NumRows()-1), 0, &val) < 0
 }
 
-// outOfRangeRowCount estimate the row count of part of [lDatum, rDatum] which is out of range of the histogram.
+// OutOfRangeRowCount estimate the row count of part of [lDatum, rDatum] which is out of range of the histogram.
 // Here we assume the density of data is decreasing from the lower/upper bound of the histogram toward outside.
 // The maximum row count it can get is the modifyCount. It reaches the maximum when out-of-range width reaches histogram range width.
 // As it shows below. To calculate the out-of-range row count, we need to calculate the percentage of the shaded area.
@@ -847,7 +849,7 @@ func (hg *Histogram) outOfRange(val types.Datum) bool {
          │   │
     lDatum  rDatum
 */
-func (hg *Histogram) outOfRangeRowCount(sctx sessionctx.Context, lDatum, rDatum *types.Datum, modifyCount int64) (result float64) {
+func (hg *Histogram) OutOfRangeRowCount(sctx sessionctx.Context, lDatum, rDatum *types.Datum, modifyCount int64) (result float64) {
 	debugTrace := sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace
 	if debugTrace {
 		debugtrace.EnterContextCommon(sctx)
@@ -965,7 +967,7 @@ func (hg *Histogram) outOfRangeRowCount(sctx sessionctx.Context, lDatum, rDatum 
 	if totalPercent > 1 {
 		totalPercent = 1
 	}
-	rowCount = totalPercent * hg.notNullCount()
+	rowCount = totalPercent * hg.NotNullCount()
 
 	// Use the modifyCount as the upper bound. Note that modifyCount contains insert, delete and update. So this is
 	// a rather loose upper bound.
@@ -1036,7 +1038,7 @@ func (hg *Histogram) ExtractTopN(cms *CMSketch, topN *TopN, numCols int, numTopN
 	dataCnts := make([]dataCnt, 0, hg.Bounds.NumRows())
 	hg.PreCalculateScalar()
 	// Set a limit on the frequency of boundary values to avoid extract values with low frequency.
-	limit := hg.notNullCount() / float64(hg.Len())
+	limit := hg.NotNullCount() / float64(hg.Len())
 	// Since our histogram are equal depth, they must occurs on the boundaries of buckets.
 	for i := 0; i < hg.Bounds.NumRows(); i++ {
 		data := hg.Bounds.GetRow(i).GetBytes(0)
@@ -1464,7 +1466,7 @@ func MergePartitionHist2GlobalHist(sc *stmtctx.StatementContext, hists []*Histog
 	for _, bucket := range globalBuckets {
 		var repeat float64
 		for _, hist := range hists {
-			histRowCount, _ := hist.equalRowCount(nil, *bucket.upper, isIndex)
+			histRowCount, _ := hist.EqualRowCount(nil, *bucket.upper, isIndex)
 			repeat += histRowCount // only hists of indexes have bucket.NDV
 		}
 		if int64(repeat) > bucket.Repeat {
