+++ statistics/merge_worker.go
@@ -166,7 +166,7 @@ func (worker *topnStatsMergeWorker) Run(timeZone *time.Location, isIndex bool,
 						datum = d
 					}
 					// Get the row count which the value is equal to the encodedVal from histogram.
-					count, _ := allHists[j].equalRowCount(nil, datum, isIndex)
+					count, _ := allHists[j].EqualRowCount(nil, datum, isIndex)
 					if count != 0 {
 						counter[encodedVal] += count
 						// Remove the value corresponding to encodedVal from the histogram.
