+++ statistics/scalar.go
@@ -17,12 +17,10 @@ package statistics
 import (
 	"encoding/binary"
 	"math"
-	"time"
 
 	"github.com/pingcap/tidb/parser/mysql"
 	"github.com/pingcap/tidb/sessionctx/stmtctx"
 	"github.com/pingcap/tidb/types"
-	"github.com/pingcap/tidb/util/mathutil"
 )
 
 // calcFraction is used to calculate the fraction of the interval [lower, upper] that lies within the [lower, value]
@@ -191,124 +189,4 @@ func calcFraction4Datums(lower, upper, value *types.Datum) float64 {
 		return calcFraction(convertDatumToScalar(lower, commonPfxLen), convertDatumToScalar(upper, commonPfxLen), convertDatumToScalar(value, commonPfxLen))
 	}
 	return 0.5
-}
-
-const maxNumStep = 10
-
-func enumRangeValues(low, high types.Datum, lowExclude, highExclude bool) []types.Datum {
-	if low.Kind() != high.Kind() {
-		return nil
-	}
-	exclude := 0
-	if lowExclude {
-		exclude++
-	}
-	if highExclude {
-		exclude++
-	}
-	switch low.Kind() {
-	case types.KindInt64:
-		// Overflow check.
-		lowVal, highVal := low.GetInt64(), high.GetInt64()
-		if lowVal <= 0 && highVal >= 0 {
-			if lowVal < -maxNumStep || highVal > maxNumStep {
-				return nil
-			}
-		}
-		remaining := highVal - lowVal
-		if remaining >= maxNumStep+1 {
-			return nil
-		}
-		remaining = remaining + 1 - int64(exclude)
-		if remaining >= maxNumStep || remaining < 0 {
-			return nil
-		}
-		values := make([]types.Datum, 0, remaining)
-		startValue := lowVal
-		if lowExclude {
-			startValue++
-		}
-		for i := int64(0); i < remaining; i++ {
-			values = append(values, types.NewIntDatum(startValue+i))
-		}
-		return values
-	case types.KindUint64:
-		remaining := high.GetUint64() - low.GetUint64()
-		if remaining >= maxNumStep+1 {
-			return nil
-		}
-		remaining = remaining + 1 - uint64(exclude)
-		if remaining >= maxNumStep || remaining < 0 {
-			return nil
-		}
-		values := make([]types.Datum, 0, remaining)
-		startValue := low.GetUint64()
-		if lowExclude {
-			startValue++
-		}
-		for i := uint64(0); i < remaining; i++ {
-			values = append(values, types.NewUintDatum(startValue+i))
-		}
-		return values
-	case types.KindMysqlDuration:
-		lowDur, highDur := low.GetMysqlDuration(), high.GetMysqlDuration()
-		fsp := mathutil.Max(lowDur.Fsp, highDur.Fsp)
-		stepSize := int64(math.Pow10(types.MaxFsp-fsp)) * int64(time.Microsecond)
-		lowDur.Duration = lowDur.Duration.Round(time.Duration(stepSize))
-		remaining := int64(highDur.Duration-lowDur.Duration)/stepSize + 1 - int64(exclude)
-		if remaining <= 0 || remaining >= maxNumStep {
-			return nil
-		}
-		startValue := int64(lowDur.Duration)
-		if lowExclude {
-			startValue += stepSize
-		}
-		values := make([]types.Datum, 0, remaining)
-		for i := int64(0); i < remaining; i++ {
-			values = append(values, types.NewDurationDatum(types.Duration{Duration: time.Duration(startValue + i*stepSize), Fsp: fsp}))
-		}
-		return values
-	case types.KindMysqlTime:
-		lowTime, highTime := low.GetMysqlTime(), high.GetMysqlTime()
-		if lowTime.Type() != highTime.Type() {
-			return nil
-		}
-		fsp := mathutil.Max(lowTime.Fsp(), highTime.Fsp())
-		var stepSize int64
-		sc := &stmtctx.StatementContext{TimeZone: time.UTC}
-		if lowTime.Type() == mysql.TypeDate {
-			stepSize = 24 * int64(time.Hour)
-			lowTime.SetCoreTime(types.FromDate(lowTime.Year(), lowTime.Month(), lowTime.Day(), 0, 0, 0, 0))
-		} else {
-			var err error
-			lowTime, err = lowTime.RoundFrac(sc, fsp)
-			if err != nil {
-				return nil
-			}
-			stepSize = int64(math.Pow10(types.MaxFsp-fsp)) * int64(time.Microsecond)
-		}
-		remaining := int64(highTime.Sub(sc, &lowTime).Duration)/stepSize + 1 - int64(exclude)
-		// When `highTime` is much larger than `lowTime`, `remaining` may be overflowed to a negative value.
-		if remaining <= 0 || remaining >= maxNumStep {
-			return nil
-		}
-		startValue := lowTime
-		var err error
-		if lowExclude {
-			startValue, err = lowTime.Add(sc, types.Duration{Duration: time.Duration(stepSize), Fsp: fsp})
-			if err != nil {
-				return nil
-			}
-		}
-		values := make([]types.Datum, 0, remaining)
-		for i := int64(0); i < remaining; i++ {
-			value, err := startValue.Add(sc, types.Duration{Duration: time.Duration(i * stepSize), Fsp: fsp})
-			if err != nil {
-				return nil
-			}
-			values = append(values, types.NewTimeDatum(value))
-		}
-		return values
-	}
-	return nil
 }
