+++ statistics/table.go
@@ -17,30 +17,23 @@ package statistics
 import (
 	"cmp"
 	"fmt"
-	"math"
 	"slices"
 	"strings"
 	"sync"
 
-	"github.com/pingcap/errors"
 	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/parser/model"
 	"github.com/pingcap/tidb/parser/mysql"
-	"github.com/pingcap/tidb/planner/util/debugtrace"
 	"github.com/pingcap/tidb/sessionctx"
-	"github.com/pingcap/tidb/sessionctx/stmtctx"
 	"github.com/pingcap/tidb/tablecodec"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util/chunk"
 	"github.com/pingcap/tidb/util/codec"
 	"github.com/pingcap/tidb/util/collate"
-	"github.com/pingcap/tidb/util/logutil"
 	"github.com/pingcap/tidb/util/mathutil"
 	"github.com/pingcap/tidb/util/ranger"
-	"github.com/pingcap/tidb/util/tracing"
 	"go.uber.org/atomic"
-	"go.uber.org/zap"
 )
 
 const (
@@ -478,228 +471,6 @@ func (t *Table) IsOutdated() bool {
 	return false
 }
 
-// ColumnGreaterRowCount estimates the row count where the column greater than value.
-func (t *Table) ColumnGreaterRowCount(sctx sessionctx.Context, value types.Datum, colID int64) float64 {
-	c, ok := t.Columns[colID]
-	if !ok || c.IsInvalid(sctx, t.Pseudo) {
-		return float64(t.RealtimeCount) / pseudoLessRate
-	}
-	return c.greaterRowCount(value) * c.GetIncreaseFactor(t.RealtimeCount)
-}
-
-// ColumnLessRowCount estimates the row count where the column less than value. Note that null values are not counted.
-func (t *Table) ColumnLessRowCount(sctx sessionctx.Context, value types.Datum, colID int64) float64 {
-	c, ok := t.Columns[colID]
-	if !ok || c.IsInvalid(sctx, t.Pseudo) {
-		return float64(t.RealtimeCount) / pseudoLessRate
-	}
-	return c.lessRowCount(sctx, value) * c.GetIncreaseFactor(t.RealtimeCount)
-}
-
-// ColumnBetweenRowCount estimates the row count where column greater or equal to a and less than b.
-func (t *Table) ColumnBetweenRowCount(sctx sessionctx.Context, a, b types.Datum, colID int64) (float64, error) {
-	sc := sctx.GetSessionVars().StmtCtx
-	c, ok := t.Columns[colID]
-	if !ok || c.IsInvalid(sctx, t.Pseudo) {
-		return float64(t.RealtimeCount) / pseudoBetweenRate, nil
-	}
-	aEncoded, err := codec.EncodeKey(sc, nil, a)
-	if err != nil {
-		return 0, err
-	}
-	bEncoded, err := codec.EncodeKey(sc, nil, b)
-	if err != nil {
-		return 0, err
-	}
-	count := c.BetweenRowCount(sctx, a, b, aEncoded, bEncoded)
-	if a.IsNull() {
-		count += float64(c.NullCount)
-	}
-	return count * c.GetIncreaseFactor(t.RealtimeCount), nil
-}
-
-// ColumnEqualRowCount estimates the row count where the column equals to value.
-func (t *Table) ColumnEqualRowCount(sctx sessionctx.Context, value types.Datum, colID int64) (float64, error) {
-	c, ok := t.Columns[colID]
-	if !ok || c.IsInvalid(sctx, t.Pseudo) {
-		return float64(t.RealtimeCount) / pseudoEqualRate, nil
-	}
-	encodedVal, err := codec.EncodeKey(sctx.GetSessionVars().StmtCtx, nil, value)
-	if err != nil {
-		return 0, err
-	}
-	result, err := c.equalRowCount(sctx, value, encodedVal, t.ModifyCount)
-	result *= c.GetIncreaseFactor(t.RealtimeCount)
-	return result, errors.Trace(err)
-}
-
-// GetRowCountByIntColumnRanges estimates the row count by a slice of IntColumnRange.
-func (coll *HistColl) GetRowCountByIntColumnRanges(sctx sessionctx.Context, colID int64, intRanges []*ranger.Range) (result float64, err error) {
-	var name string
-	if sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
-		debugtrace.EnterContextCommon(sctx)
-		debugTraceGetRowCountInput(sctx, colID, intRanges)
-		defer func() {
-			debugtrace.RecordAnyValuesWithNames(sctx, "Name", name, "Result", result)
-			debugtrace.LeaveContextCommon(sctx)
-		}()
-	}
-	sc := sctx.GetSessionVars().StmtCtx
-	c, ok := coll.Columns[colID]
-	recordUsedItemStatsStatus(sctx, c, coll.PhysicalID, colID)
-	if c != nil && c.Info != nil {
-		name = c.Info.Name.O
-	}
-	if !ok || c.IsInvalid(sctx, coll.Pseudo) {
-		if len(intRanges) == 0 {
-			return 0, nil
-		}
-		if intRanges[0].LowVal[0].Kind() == types.KindInt64 {
-			result = getPseudoRowCountBySignedIntRanges(intRanges, float64(coll.RealtimeCount))
-		} else {
-			result = getPseudoRowCountByUnsignedIntRanges(intRanges, float64(coll.RealtimeCount))
-		}
-		if sc.EnableOptimizerCETrace && ok {
-			CETraceRange(sctx, coll.PhysicalID, []string{c.Info.Name.O}, intRanges, "Column Stats-Pseudo", uint64(result))
-		}
-		return result, nil
-	}
-	if sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
-		debugtrace.RecordAnyValuesWithNames(sctx,
-			"Histogram NotNull Count", c.Histogram.notNullCount(),
-			"TopN total count", c.TopN.TotalCount(),
-			"Increase Factor", c.GetIncreaseFactor(coll.RealtimeCount),
-		)
-	}
-	result, err = c.GetColumnRowCount(sctx, intRanges, coll.RealtimeCount, coll.ModifyCount, true)
-	if sc.EnableOptimizerCETrace {
-		CETraceRange(sctx, coll.PhysicalID, []string{c.Info.Name.O}, intRanges, "Column Stats", uint64(result))
-	}
-	return result, errors.Trace(err)
-}
-
-// GetRowCountByColumnRanges estimates the row count by a slice of Range.
-func (coll *HistColl) GetRowCountByColumnRanges(sctx sessionctx.Context, colID int64, colRanges []*ranger.Range) (result float64, err error) {
-	var name string
-	if sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
-		debugtrace.EnterContextCommon(sctx)
-		debugTraceGetRowCountInput(sctx, colID, colRanges)
-		defer func() {
-			debugtrace.RecordAnyValuesWithNames(sctx, "Name", name, "Result", result)
-			debugtrace.LeaveContextCommon(sctx)
-		}()
-	}
-	sc := sctx.GetSessionVars().StmtCtx
-	c, ok := coll.Columns[colID]
-	recordUsedItemStatsStatus(sctx, c, coll.PhysicalID, colID)
-	if c != nil && c.Info != nil {
-		name = c.Info.Name.O
-	}
-	if !ok || c.IsInvalid(sctx, coll.Pseudo) {
-		result, err = GetPseudoRowCountByColumnRanges(sc, float64(coll.RealtimeCount), colRanges, 0)
-		if err == nil && sc.EnableOptimizerCETrace && ok {
-			CETraceRange(sctx, coll.PhysicalID, []string{c.Info.Name.O}, colRanges, "Column Stats-Pseudo", uint64(result))
-		}
-		return result, err
-	}
-	if sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
-		debugtrace.RecordAnyValuesWithNames(sctx,
-			"Histogram NotNull Count", c.Histogram.notNullCount(),
-			"TopN total count", c.TopN.TotalCount(),
-			"Increase Factor", c.GetIncreaseFactor(coll.RealtimeCount),
-		)
-	}
-	result, err = c.GetColumnRowCount(sctx, colRanges, coll.RealtimeCount, coll.ModifyCount, false)
-	if sc.EnableOptimizerCETrace {
-		CETraceRange(sctx, coll.PhysicalID, []string{c.Info.Name.O}, colRanges, "Column Stats", uint64(result))
-	}
-	return result, errors.Trace(err)
-}
-
-// GetRowCountByIndexRanges estimates the row count by a slice of Range.
-func (coll *HistColl) GetRowCountByIndexRanges(sctx sessionctx.Context, idxID int64, indexRanges []*ranger.Range) (result float64, err error) {
-	var name string
-	if sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
-		debugtrace.EnterContextCommon(sctx)
-		debugTraceGetRowCountInput(sctx, idxID, indexRanges)
-		defer func() {
-			debugtrace.RecordAnyValuesWithNames(sctx, "Name", name, "Result", result)
-			debugtrace.LeaveContextCommon(sctx)
-		}()
-	}
-	sc := sctx.GetSessionVars().StmtCtx
-	idx, ok := coll.Indices[idxID]
-	colNames := make([]string, 0, 8)
-	if ok {
-		if idx.Info != nil {
-			name = idx.Info.Name.O
-			for _, col := range idx.Info.Columns {
-				colNames = append(colNames, col.Name.O)
-			}
-		}
-	}
-	recordUsedItemStatsStatus(sctx, idx, coll.PhysicalID, idxID)
-	if !ok || idx.IsInvalid(sctx, coll.Pseudo) {
-		colsLen := -1
-		if idx != nil && idx.Info.Unique {
-			colsLen = len(idx.Info.Columns)
-		}
-		result, err = getPseudoRowCountByIndexRanges(sc, indexRanges, float64(coll.RealtimeCount), colsLen)
-		if err == nil && sc.EnableOptimizerCETrace && ok {
-			CETraceRange(sctx, coll.PhysicalID, colNames, indexRanges, "Index Stats-Pseudo", uint64(result))
-		}
-		return result, err
-	}
-	if sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
-		debugtrace.RecordAnyValuesWithNames(sctx,
-			"Histogram NotNull Count", idx.Histogram.notNullCount(),
-			"TopN total count", idx.TopN.TotalCount(),
-			"Increase Factor", idx.GetIncreaseFactor(coll.RealtimeCount),
-		)
-	}
-	if idx.CMSketch != nil && idx.StatsVer == Version1 {
-		result, err = coll.getIndexRowCount(sctx, idxID, indexRanges)
-	} else {
-		result, err = idx.GetRowCount(sctx, coll, indexRanges, coll.RealtimeCount, coll.ModifyCount)
-	}
-	if sc.EnableOptimizerCETrace {
-		CETraceRange(sctx, coll.PhysicalID, colNames, indexRanges, "Index Stats", uint64(result))
-	}
-	return result, errors.Trace(err)
-}
-
-// CETraceRange appends a list of ranges and related information into CE trace
-func CETraceRange(sctx sessionctx.Context, tableID int64, colNames []string, ranges []*ranger.Range, tp string, rowCount uint64) {
-	sc := sctx.GetSessionVars().StmtCtx
-	allPoint := true
-	for _, ran := range ranges {
-		if !ran.IsPointNullable(sctx) {
-			allPoint = false
-			break
-		}
-	}
-	if allPoint {
-		tp = tp + "-Point"
-	} else {
-		tp = tp + "-Range"
-	}
-	expr, err := ranger.RangesToString(sc, ranges, colNames)
-	if err != nil {
-		logutil.BgLogger().Debug("Failed to trace CE of ranges", zap.String("category", "OptimizerTrace"), zap.Error(err))
-	}
-	// We don't need to record meaningless expressions.
-	if expr == "" || expr == "true" || expr == "false" {
-		return
-	}
-	ceRecord := tracing.CETraceRecord{
-		TableID:  tableID,
-		Type:     tp,
-		Expr:     expr,
-		RowCount: rowCount,
-	}
-	sc.OptimizerCETrace = append(sc.OptimizerCETrace, &ceRecord)
-}
-
 func (coll *HistColl) findAvailableStatsForCol(sctx sessionctx.Context, uniqueID int64) (isIndex bool, idx int64) {
 	// try to find available stats in column stats
 	if colStats, ok := coll.Columns[uniqueID]; ok && colStats != nil && !colStats.IsInvalid(sctx, coll.Pseudo) && colStats.IsFullLoad() {
@@ -773,7 +544,7 @@ func (coll *HistColl) GetSelectivityByFilter(sctx sessionctx.Context, filters []
 		return false, 0, nil
 	}
 	topnTotalCnt = topn.TotalCount()
-	histTotalCnt = hist.notNullCount()
+	histTotalCnt = hist.NotNullCount()
 	totalCnt = float64(topnTotalCnt) + histTotalCnt + float64(nullCnt)
 
 	var topNSel, histSel, nullSel float64
@@ -875,22 +646,6 @@ func (t *Table) PseudoAvgCountPerValue() float64 {
 	return float64(t.RealtimeCount) / pseudoEqualRate
 }
 
-// GetOrdinalOfRangeCond gets the ordinal of the position range condition,
-// if not exist, it returns the end position.
-func GetOrdinalOfRangeCond(sc *stmtctx.StatementContext, ran *ranger.Range) int {
-	for i := range ran.LowVal {
-		a, b := ran.LowVal[i], ran.HighVal[i]
-		cmp, err := a.Compare(sc, &b, ran.Collators[0])
-		if err != nil {
-			return 0
-		}
-		if cmp != 0 {
-			return i
-		}
-	}
-	return len(ran.LowVal)
-}
-
 // ID2UniqueID generates a new HistColl whose `Columns` is built from UniqueID of given columns.
 func (coll *HistColl) ID2UniqueID(columns []*expression.Column) *HistColl {
 	cols := make(map[int64]*Column)
@@ -970,276 +725,6 @@ func (coll *HistColl) GenerateHistCollFromColumnInfo(tblInfo *model.TableInfo, c
 	return newColl
 }
 
-// isSingleColIdxNullRange checks if a range is [NULL, NULL] on a single-column index.
-func isSingleColIdxNullRange(idx *Index, ran *ranger.Range) bool {
-	if len(idx.Info.Columns) > 1 {
-		return false
-	}
-	l, h := ran.LowVal[0], ran.HighVal[0]
-	if l.IsNull() && h.IsNull() {
-		return true
-	}
-	return false
-}
-
-// outOfRangeEQSelectivity estimates selectivities for out-of-range values.
-// It assumes all modifications are insertions and all new-inserted rows are uniformly distributed
-// and has the same distribution with analyzed rows, which means each unique value should have the
-// same number of rows(Tot/NDV) of it.
-// The input sctx is just for debug trace, you can pass nil safely if that's not needed.
-func outOfRangeEQSelectivity(sctx sessionctx.Context, ndv, realtimeRowCount, columnRowCount int64) (result float64) {
-	if sctx != nil && sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
-		debugtrace.EnterContextCommon(sctx)
-		defer func() {
-			debugtrace.RecordAnyValuesWithNames(sctx, "Result", result)
-			debugtrace.LeaveContextCommon(sctx)
-		}()
-	}
-	increaseRowCount := realtimeRowCount - columnRowCount
-	if increaseRowCount <= 0 {
-		return 0 // it must be 0 since the histogram contains the whole data
-	}
-	if ndv < outOfRangeBetweenRate {
-		ndv = outOfRangeBetweenRate // avoid inaccurate selectivity caused by small NDV
-	}
-	selectivity := 1 / float64(ndv)
-	if selectivity*float64(columnRowCount) > float64(increaseRowCount) {
-		selectivity = float64(increaseRowCount) / float64(columnRowCount)
-	}
-	return selectivity
-}
-
-// crossValidationSelectivity gets the selectivity of multi-column equal conditions by cross validation.
-func (coll *HistColl) crossValidationSelectivity(
-	sctx sessionctx.Context,
-	idx *Index,
-	usedColsLen int,
-	idxPointRange *ranger.Range,
-) (
-	minRowCount float64,
-	crossValidationSelectivity float64,
-	err error,
-) {
-	if sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
-		debugtrace.EnterContextCommon(sctx)
-		defer func() {
-			var idxName string
-			if idx != nil && idx.Info != nil {
-				idxName = idx.Info.Name.O
-			}
-			debugtrace.RecordAnyValuesWithNames(
-				sctx,
-				"Index Name", idxName,
-				"minRowCount", minRowCount,
-				"crossValidationSelectivity", crossValidationSelectivity,
-				"error", err,
-			)
-			debugtrace.LeaveContextCommon(sctx)
-		}()
-	}
-	minRowCount = math.MaxFloat64
-	cols := coll.Idx2ColumnIDs[idx.ID]
-	crossValidationSelectivity = 1.0
-	totalRowCount := idx.TotalRowCount()
-	for i, colID := range cols {
-		if i >= usedColsLen {
-			break
-		}
-		if col, ok := coll.Columns[colID]; ok {
-			if col.IsInvalid(sctx, coll.Pseudo) {
-				continue
-			}
-			// Since the column range is point range(LowVal is equal to HighVal), we need to set both LowExclude and HighExclude to false.
-			// Otherwise we would get 0.0 estRow from GetColumnRowCount.
-			rang := ranger.Range{
-				LowVal:      []types.Datum{idxPointRange.LowVal[i]},
-				LowExclude:  false,
-				HighVal:     []types.Datum{idxPointRange.HighVal[i]},
-				HighExclude: false,
-				Collators:   []collate.Collator{idxPointRange.Collators[i]},
-			}
-
-			rowCount, err := col.GetColumnRowCount(sctx, []*ranger.Range{&rang}, coll.RealtimeCount, coll.ModifyCount, col.IsHandle)
-			if err != nil {
-				return 0, 0, err
-			}
-			crossValidationSelectivity = crossValidationSelectivity * (rowCount / totalRowCount)
-
-			if rowCount < minRowCount {
-				minRowCount = rowCount
-			}
-		}
-	}
-	return minRowCount, crossValidationSelectivity, nil
-}
-
-// getEqualCondSelectivity gets the selectivity of the equal conditions.
-func (coll *HistColl) getEqualCondSelectivity(sctx sessionctx.Context, idx *Index, bytes []byte, usedColsLen int, idxPointRange *ranger.Range) (result float64, err error) {
-	if sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
-		debugtrace.EnterContextCommon(sctx)
-		defer func() {
-			var idxName string
-			if idx != nil && idx.Info != nil {
-				idxName = idx.Info.Name.O
-			}
-			debugtrace.RecordAnyValuesWithNames(
-				sctx,
-				"Index Name", idxName,
-				"Encoded", bytes,
-				"UsedColLen", usedColsLen,
-				"Range", idxPointRange.String(),
-				"Result", result,
-				"error", err,
-			)
-			debugtrace.LeaveContextCommon(sctx)
-		}()
-	}
-	coverAll := len(idx.Info.Columns) == usedColsLen
-	// In this case, the row count is at most 1.
-	if idx.Info.Unique && coverAll {
-		return 1.0 / idx.TotalRowCount(), nil
-	}
-	val := types.NewBytesDatum(bytes)
-	if idx.outOfRange(val) {
-		// When the value is out of range, we could not found this value in the CM Sketch,
-		// so we use heuristic methods to estimate the selectivity.
-		if idx.NDV > 0 && coverAll {
-			return outOfRangeEQSelectivity(sctx, idx.NDV, coll.RealtimeCount, int64(idx.TotalRowCount())), nil
-		}
-		// The equal condition only uses prefix columns of the index.
-		colIDs := coll.Idx2ColumnIDs[idx.ID]
-		var ndv int64
-		for i, colID := range colIDs {
-			if i >= usedColsLen {
-				break
-			}
-			if col, ok := coll.Columns[colID]; ok {
-				ndv = mathutil.Max(ndv, col.Histogram.NDV)
-			}
-		}
-		return outOfRangeEQSelectivity(sctx, ndv, coll.RealtimeCount, int64(idx.TotalRowCount())), nil
-	}
-
-	minRowCount, crossValidationSelectivity, err := coll.crossValidationSelectivity(sctx, idx, usedColsLen, idxPointRange)
-	if err != nil {
-		return 0, err
-	}
-
-	idxCount := float64(idx.QueryBytes(sctx, bytes))
-	if minRowCount < idxCount {
-		return crossValidationSelectivity, nil
-	}
-	return idxCount / idx.TotalRowCount(), nil
-}
-
-func (coll *HistColl) getIndexRowCount(sctx sessionctx.Context, idxID int64, indexRanges []*ranger.Range) (float64, error) {
-	sc := sctx.GetSessionVars().StmtCtx
-	debugTrace := sc.EnableOptimizerDebugTrace
-	if debugTrace {
-		debugtrace.EnterContextCommon(sctx)
-		defer debugtrace.LeaveContextCommon(sctx)
-	}
-	idx := coll.Indices[idxID]
-	totalCount := float64(0)
-	for _, ran := range indexRanges {
-		if debugTrace {
-			debugTraceStartEstimateRange(sctx, ran, nil, nil, totalCount)
-		}
-		rangePosition := GetOrdinalOfRangeCond(sc, ran)
-		var rangeVals []types.Datum
-		// Try to enum the last range values.
-		if rangePosition != len(ran.LowVal) {
-			rangeVals = enumRangeValues(ran.LowVal[rangePosition], ran.HighVal[rangePosition], ran.LowExclude, ran.HighExclude)
-			if rangeVals != nil {
-				rangePosition++
-			}
-		}
-		// If first one is range, just use the previous way to estimate; if it is [NULL, NULL] range
-		// on single-column index, use previous way as well, because CMSketch does not contain null
-		// values in this case.
-		if rangePosition == 0 || isSingleColIdxNullRange(idx, ran) {
-			count, err := idx.GetRowCount(sctx, nil, []*ranger.Range{ran}, coll.RealtimeCount, coll.ModifyCount)
-			if err != nil {
-				return 0, errors.Trace(err)
-			}
-			if debugTrace {
-				debugTraceEndEstimateRange(sctx, count, debugTraceRange)
-			}
-			totalCount += count
-			continue
-		}
-		var selectivity float64
-		// use CM Sketch to estimate the equal conditions
-		if rangeVals == nil {
-			bytes, err := codec.EncodeKey(sc, nil, ran.LowVal[:rangePosition]...)
-			if err != nil {
-				return 0, errors.Trace(err)
-			}
-			selectivity, err = coll.getEqualCondSelectivity(sctx, idx, bytes, rangePosition, ran)
-			if err != nil {
-				return 0, errors.Trace(err)
-			}
-		} else {
-			bytes, err := codec.EncodeKey(sc, nil, ran.LowVal[:rangePosition-1]...)
-			if err != nil {
-				return 0, errors.Trace(err)
-			}
-			prefixLen := len(bytes)
-			for _, val := range rangeVals {
-				bytes = bytes[:prefixLen]
-				bytes, err = codec.EncodeKey(sc, bytes, val)
-				if err != nil {
-					return 0, err
-				}
-				res, err := coll.getEqualCondSelectivity(sctx, idx, bytes, rangePosition, ran)
-				if err != nil {
-					return 0, errors.Trace(err)
-				}
-				selectivity += res
-			}
-		}
-		// use histogram to estimate the range condition
-		if rangePosition != len(ran.LowVal) {
-			rang := ranger.Range{
-				LowVal:      []types.Datum{ran.LowVal[rangePosition]},
-				LowExclude:  ran.LowExclude,
-				HighVal:     []types.Datum{ran.HighVal[rangePosition]},
-				HighExclude: ran.HighExclude,
-				Collators:   []collate.Collator{ran.Collators[rangePosition]},
-			}
-			var count float64
-			var err error
-			colIDs := coll.Idx2ColumnIDs[idxID]
-			var colID int64
-			if rangePosition >= len(colIDs) {
-				colID = -1
-			} else {
-				colID = colIDs[rangePosition]
-			}
-			// prefer index stats over column stats
-			if idxIDs, ok := coll.ColID2IdxIDs[colID]; ok && len(idxIDs) > 0 {
-				idxID := idxIDs[0]
-				count, err = coll.GetRowCountByIndexRanges(sctx, idxID, []*ranger.Range{&rang})
-			} else {
-				count, err = coll.GetRowCountByColumnRanges(sctx, colID, []*ranger.Range{&rang})
-			}
-			if err != nil {
-				return 0, errors.Trace(err)
-			}
-			selectivity = selectivity * count / idx.TotalRowCount()
-		}
-		count := selectivity * idx.TotalRowCount()
-		if debugTrace {
-			debugTraceEndEstimateRange(sctx, count, debugTraceRange)
-		}
-		totalCount += count
-	}
-	if totalCount > idx.TotalRowCount() {
-		totalCount = idx.TotalRowCount()
-	}
-	return totalCount, nil
-}
-
 // PseudoTable creates a pseudo table statistics.
 func PseudoTable(tblInfo *model.TableInfo) *Table {
 	const fakePhysicalID int64 = -1
@@ -1277,149 +762,6 @@ func PseudoTable(tblInfo *model.TableInfo) *Table {
 		}
 	}
 	return t
-}
-
-func getPseudoRowCountByIndexRanges(sc *stmtctx.StatementContext, indexRanges []*ranger.Range,
-	tableRowCount float64, colsLen int) (float64, error) {
-	if tableRowCount == 0 {
-		return 0, nil
-	}
-	var totalCount float64
-	for _, indexRange := range indexRanges {
-		count := tableRowCount
-		i, err := indexRange.PrefixEqualLen(sc)
-		if err != nil {
-			return 0, errors.Trace(err)
-		}
-		if i == colsLen && !indexRange.LowExclude && !indexRange.HighExclude {
-			totalCount += 1.0
-			continue
-		}
-		if i >= len(indexRange.LowVal) {
-			i = len(indexRange.LowVal) - 1
-		}
-		rowCount, err := GetPseudoRowCountByColumnRanges(sc, tableRowCount, []*ranger.Range{indexRange}, i)
-		if err != nil {
-			return 0, errors.Trace(err)
-		}
-		count = count / tableRowCount * rowCount
-		// If the condition is a = 1, b = 1, c = 1, d = 1, we think every a=1, b=1, c=1 only filtrate 1/100 data,
-		// so as to avoid collapsing too fast.
-		for j := 0; j < i; j++ {
-			count = count / float64(100)
-		}
-		totalCount += count
-	}
-	if totalCount > tableRowCount {
-		totalCount = tableRowCount / 3.0
-	}
-	return totalCount, nil
-}
-
-// GetPseudoRowCountByColumnRanges calculate the row count by the ranges if there's no statistics information for this column.
-func GetPseudoRowCountByColumnRanges(sc *stmtctx.StatementContext, tableRowCount float64, columnRanges []*ranger.Range, colIdx int) (float64, error) {
-	var rowCount float64
-	for _, ran := range columnRanges {
-		if ran.LowVal[colIdx].Kind() == types.KindNull && ran.HighVal[colIdx].Kind() == types.KindMaxValue {
-			rowCount += tableRowCount
-		} else if ran.LowVal[colIdx].Kind() == types.KindMinNotNull {
-			nullCount := tableRowCount / pseudoEqualRate
-			if ran.HighVal[colIdx].Kind() == types.KindMaxValue {
-				rowCount += tableRowCount - nullCount
-			} else {
-				lessCount := tableRowCount / pseudoLessRate
-				rowCount += lessCount - nullCount
-			}
-		} else if ran.HighVal[colIdx].Kind() == types.KindMaxValue {
-			rowCount += tableRowCount / pseudoLessRate
-		} else {
-			compare, err := ran.LowVal[colIdx].Compare(sc, &ran.HighVal[colIdx], ran.Collators[colIdx])
-			if err != nil {
-				return 0, errors.Trace(err)
-			}
-			if compare == 0 {
-				rowCount += tableRowCount / pseudoEqualRate
-			} else {
-				rowCount += tableRowCount / pseudoBetweenRate
-			}
-		}
-	}
-	if rowCount > tableRowCount {
-		rowCount = tableRowCount
-	}
-	return rowCount, nil
-}
-
-func getPseudoRowCountBySignedIntRanges(intRanges []*ranger.Range, tableRowCount float64) float64 {
-	var rowCount float64
-	for _, rg := range intRanges {
-		var cnt float64
-		low := rg.LowVal[0].GetInt64()
-		if rg.LowVal[0].Kind() == types.KindNull || rg.LowVal[0].Kind() == types.KindMinNotNull {
-			low = math.MinInt64
-		}
-		high := rg.HighVal[0].GetInt64()
-		if rg.HighVal[0].Kind() == types.KindMaxValue {
-			high = math.MaxInt64
-		}
-		if low == math.MinInt64 && high == math.MaxInt64 {
-			cnt = tableRowCount
-		} else if low == math.MinInt64 {
-			cnt = tableRowCount / pseudoLessRate
-		} else if high == math.MaxInt64 {
-			cnt = tableRowCount / pseudoLessRate
-		} else {
-			if low == high {
-				cnt = 1 // When primary key is handle, the equal row count is at most one.
-			} else {
-				cnt = tableRowCount / pseudoBetweenRate
-			}
-		}
-		if high-low > 0 && cnt > float64(high-low) {
-			cnt = float64(high - low)
-		}
-		rowCount += cnt
-	}
-	if rowCount > tableRowCount {
-		rowCount = tableRowCount
-	}
-	return rowCount
-}
-
-func getPseudoRowCountByUnsignedIntRanges(intRanges []*ranger.Range, tableRowCount float64) float64 {
-	var rowCount float64
-	for _, rg := range intRanges {
-		var cnt float64
-		low := rg.LowVal[0].GetUint64()
-		if rg.LowVal[0].Kind() == types.KindNull || rg.LowVal[0].Kind() == types.KindMinNotNull {
-			low = 0
-		}
-		high := rg.HighVal[0].GetUint64()
-		if rg.HighVal[0].Kind() == types.KindMaxValue {
-			high = math.MaxUint64
-		}
-		if low == 0 && high == math.MaxUint64 {
-			cnt = tableRowCount
-		} else if low == 0 {
-			cnt = tableRowCount / pseudoLessRate
-		} else if high == math.MaxUint64 {
-			cnt = tableRowCount / pseudoLessRate
-		} else {
-			if low == high {
-				cnt = 1 // When primary key is handle, the equal row count is at most one.
-			} else {
-				cnt = tableRowCount / pseudoBetweenRate
-			}
-		}
-		if high > low && cnt > float64(high-low) {
-			cnt = float64(high - low)
-		}
-		rowCount += cnt
-	}
-	if rowCount > tableRowCount {
-		rowCount = tableRowCount
-	}
-	return rowCount
 }
 
 // GetAvgRowSize computes average row size for given columns.
@@ -1534,66 +876,16 @@ func CheckAnalyzeVerOnTable(tbl *Table, version *int) bool {
 	return true
 }
 
-// GetTblInfoForUsedStatsByPhysicalID get table name, partition name and TableInfo that will be used to record used stats.
-var GetTblInfoForUsedStatsByPhysicalID func(sctx sessionctx.Context, id int64) (fullName string, tblInfo *model.TableInfo)
+var (
+	// Below functions are used to solve cycle import problem in statistics_test.go.
+	// TODO: solve this problem in a better way.
 
-// recordUsedItemStatsStatus only records un-FullLoad item load status during user query
-func recordUsedItemStatsStatus(sctx sessionctx.Context, stats interface{}, tableID, id int64) {
-	// Sometimes we try to use stats on _tidb_rowid (id == -1), which must be empty, we ignore this case here.
-	if id <= 0 {
-		return
-	}
-	var isIndex, missing bool
-	var loadStatus *StatsLoadedStatus
-	switch x := stats.(type) {
-	case *Column:
-		isIndex = false
-		if x == nil {
-			missing = true
-		} else {
-			loadStatus = &x.StatsLoadedStatus
-		}
-	case *Index:
-		isIndex = true
-		if x == nil {
-			missing = true
-		} else {
-			loadStatus = &x.StatsLoadedStatus
-		}
-	}
+	// GetRowCountByColumnRanges is a hook for getting row count by column ranges.
+	GetRowCountByColumnRanges func(sctx sessionctx.Context, coll *HistColl, colID int64, colRanges []*ranger.Range) (result float64, err error)
 
-	// no need to record
-	if !missing && loadStatus.IsFullLoad() {
-		return
-	}
+	// GetRowCountByIntColumnRanges is a hook for getting row count by int column ranges.
+	GetRowCountByIntColumnRanges func(sctx sessionctx.Context, coll *HistColl, colID int64, intRanges []*ranger.Range) (result float64, err error)
 
-	// need to record
-	statsRecord := sctx.GetSessionVars().StmtCtx.GetUsedStatsInfo(true)
-	if statsRecord[tableID] == nil {
-		name, tblInfo := GetTblInfoForUsedStatsByPhysicalID(sctx, tableID)
-		statsRecord[tableID] = &stmtctx.UsedStatsInfoForTable{
-			Name:    name,
-			TblInfo: tblInfo,
-		}
-	}
-	recordForTbl := statsRecord[tableID]
-
-	var recordForColOrIdx map[int64]string
-	if isIndex {
-		if recordForTbl.IndexStatsLoadStatus == nil {
-			recordForTbl.IndexStatsLoadStatus = make(map[int64]string, 1)
-		}
-		recordForColOrIdx = recordForTbl.IndexStatsLoadStatus
-	} else {
-		if recordForTbl.ColumnStatsLoadStatus == nil {
-			recordForTbl.ColumnStatsLoadStatus = make(map[int64]string, 1)
-		}
-		recordForColOrIdx = recordForTbl.ColumnStatsLoadStatus
-	}
-
-	if missing {
-		recordForColOrIdx[id] = "missing"
-		return
-	}
-	recordForColOrIdx[id] = loadStatus.StatusToString()
-}
+	// GetRowCountByIndexRanges is a hook for getting row count by index ranges.
+	GetRowCountByIndexRanges func(sctx sessionctx.Context, coll *HistColl, idxID int64, indexRanges []*ranger.Range) (result float64, err error)
+)
